# openapi.yml
openapi: 3.0.1

info:
  title: Task tracker Client service API
  version: 1.0.0

paths:
  /register:
    post:
      parameters:
        - name: login
          in: query
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Registration successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: 'Request data is invalid'
        '403':
          description: 'Login already used'
        '500':
          description: 'Internal server error'
  /auth:
    post:
      parameters:
        - name: login
          in: query
          required: true
          schema:
            type: string
        - name: password
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Authorization successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: 'Request data is invalid'
        '404':
          description: 'Login or password is incorrent'
        '500':
          description: 'Internal server error'
  /update:
    put:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
                birthdate:
                  type: string
                email:
                  type: string
                phone_number:
                  type: string
      responses:
        '200':
          description: 'Update successful'
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User not found'
        '500':
          description: 'Internal server error'
  /task:
    post:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                deadline_date:
                  type: string
      responses:
        '200':
          description: 'Create task successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: integer
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User not found'
        '500':
          description: 'Internal server error'
    put:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: task_id
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                deadline_date:
                  type: string
                completion_status:
                  type: string
      responses:
        '200':
          description: 'Update task successful'
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User or task not found'
        '500':
          description: 'Internal server error'
    delete:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: task_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Delete task successful'
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User or task not found'
        '500':
          description: 'Internal server error'
    get:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: task_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Get task successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: integer
                  author_id:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
                  deadline_date:
                    type: string
                  creation_date:
                    type: string
                  completion_status:
                    type: string
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User or task not found'
        '500':
          description: 'Internal server error'
  /tasks:
    get:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: author_login
          in: query
          required: true
          schema:
            type: string
        - name: page_index
          in: query
          required: true
          schema:
            type: integer
        - name: tasks_per_page
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Get tasks page successful'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: integer
                    author_id:
                      type: integer
                    name:
                      type: string
                    description:
                      type: string
                    deadline_date:
                      type: string
                    creation_date:
                      type: string
                    completion_status:
                      type: string
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User or page not found'
        '500':
          description: 'Internal server error'
  /task/view:
    post:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: task_id
          in: query
          required: true
          schema:
            type: integer
        - name: author_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Add view task successful'
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User or task not found'
        '500':
          description: 'Internal server error'
  /task/like:
    post:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: task_id
          in: query
          required: true
          schema:
            type: integer
        - name: author_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Add like task successful'
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'User or task not found'
        '500':
          description: 'Internal server error'
  /task/stat:
    get:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: task_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Get task stat successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: integer
                  view_count:
                    type: integer
                  like_count:
                    type: integer
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '404':
          description: 'Task not found'
        '500':
          description: 'Internal server error'
  /top/tasks:
    get:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
        - name: stat_type
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Get top tasks successful'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    task_id:
                      type: integer
                    author_id:
                      type: integer
                    author_login:
                      type: string
                    view_count:
                      type: integer
                    like_count:
                      type: integer
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '500':
          description: 'Internal server error'
  /top/authors:
    get:
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Get top authors successful'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    author_id:
                      type: integer
                    author_login:
                      type: string
                    like_count:
                      type: integer
        '400':
          description: 'Request data is invalid'
        '401':
          description: 'Unauthorized'
        '500':
          description: 'Internal server error'